% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuutils-package.R
\docType{package}
\name{cuutils-package}
\alias{cuutils-package}
\title{Bits and Pieces I Didn't Find a Serious Place to Put to}
\description{
Well, different stuff. Nothing really interesting in particular,
but hopefully helpful helpers.
}
\author{
Andreas Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
\keyword{package}
