% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_files}
\alias{find_files}
\title{find files on disk}
\usage{
find_files(file_names = NA, path = ".",
  pattern = ".*\\\\.[RrSs]$|.*\\\\.[RrSs]nw$", all.files = TRUE,
  recursive = TRUE, ignore.case = FALSE)
}
\arguments{
\item{file_names}{character vector of file names (to be checked if the files
exist).}

\item{path}{see \code{\link{list.files}}.}

\item{pattern}{see \code{\link{list.files}}.}

\item{all.files}{see \code{\link{list.files}}.}

\item{recursive}{see \code{\link{list.files}}.}

\item{ignore.case}{see \code{\link{list.files}}.}
}
\value{
a character vector of file names.
}
\description{
look for files on disk, either scanning a vector of names or searching for
files with \code{\link{list.files}} and throw an error if no files are found.
}
\note{
This is merely a wrapper around \code{\link{file.exists}} or 
      \code{\link{list.files}}, depending on wheter file_names is given.
}

