% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_R_CMD_BATCH_script_path}
\alias{get_R_CMD_BATCH_script_path}
\title{get the path of the R code file in case of an R CMD BATCH run.}
\usage{
get_R_CMD_BATCH_script_path()
}
\value{
a vector of \code{\link{mode}} character giving the name of the R
code file. Will be character(0) if not in an R CMD BATCH run.
}
\description{
get the path of the R code file in case of an R CMD BATCH run.
}
\examples{
get_R_CMD_BATCH_script_path()
}
