% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_script_name}
\alias{get_script_name}
\title{get the name of the R code file or set it to \code{default}.}
\usage{
get_script_name(default = "interactive_R_session")
}
\arguments{
\item{default}{the name to return if R is run interactively.}
}
\value{
a vector of \code{\link{length}} 1 and \code{\link{mode}} 
character giving the name of the R code file if R was run via R CMD BATCH or
Rscript, the given default otherwise.
}
\description{
The code file name is retrieved only for R CMD BATCH and Rscript,
if R is used interactively, the name is set to \code{default}, 
even if you're working with code stored in a (named) file on disk.
}
\examples{
get_script_name(default = 'foobar.R')
}

