% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{get_script_path}
\alias{get_script_path}
\title{get the path of the R code file.}
\usage{
get_script_path()
}
\value{
a vector of \code{\link{length}} 1 and \code{\link{mode}} 
character giving the name of the R code file if R was run via R CMD BATCH or
Rscript.
}
\description{
This is just a wrapper for \code{\link{get_Rscript_script_path}} and 
\code{\link{get_R_CMD_BATCH_script_path}}.
}
\examples{
get_script_path()
}

