% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_ration.R
\name{golden_ratio}
\alias{golden_ratio}
\title{Calculate the Golden Ratio}
\usage{
golden_ratio(value = 1, quantity = "a + b")
}
\arguments{
\item{value}{The value of the quantity.}

\item{quantity}{The quantity of which the value is known.}
}
\value{
a numeric vector of lenght 2, containing the two quantites a and b, 
(a being the larger) and their sum.
}
\description{
Calculate the Golden Ratio
}
\examples{
golden_ratio(10)
golden_ratio(golden_ratio(10)[["a"]], "a + b")
}
