% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_github.R
\name{install_github}
\alias{install_github}
\title{Install Packages From Github at FVAFR}
\usage{
install_github(github_repo, ignore_ostype = TRUE, ignore_r_version = TRUE,
  force = TRUE)
}
\arguments{
\item{github_repo}{A string of form "user/repo", "hadley/devtools" for
example.}

\item{ignore_ostype}{Ignore operation system type required in the
DESCRIPTION?}

\item{ignore_r_version}{Ignore R version required in the DESCRIPTION?}

\item{force}{Force even if not on windows?}
}
\value{
The value of \code{\link[utils:install.packages]{install.packages}}.
}
\description{
At fvafr curl under windows doesn't work properly. Maybe the proxy isn't set
properly. So this functions uses
\code{\link[utils:download.file]{download.file}}, which works.
}
