% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_32bit}
\alias{is_32bit}
\title{Is the CPU Architecture R was build for 32 bit?}
\usage{
is_32bit()
}
\value{
\code{\link{TRUE}} if so, \code{\link{FALSE}} otherwise.
}
\description{
Is the CPU Architecture R was build for 32 bit?
}
