% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{is_batch}
\alias{is_batch}
\title{is R run in batch mode (via R CMD BATCH or Rscript)?}
\usage{
is_batch()
}
\value{
a boolean.
}
\description{
is R run in batch mode (via R CMD BATCH or Rscript)?
}
\examples{
is_batch()
}
