% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is_windows}
\alias{is_windows}
\title{Is the System Running a Windows Machine?}
\usage{
is_windows()
}
\value{
\code{\link{TRUE}} if so, \code{\link{FALSE}} otherwise.
}
\description{
Is the System Running a Windows Machine?
}
