% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch.R
\name{provide_output_directory}
\alias{provide_output_directory}
\title{provide an output directory and a variable containing its name.}
\usage{
provide_output_directory(type = "graphics", path = ".")
}
\arguments{
\item{type}{a character vector of length 1 giving the a name postfix.
Defaults to 'graphics'.}

\item{path}{a \code{\link{file.path}}, defaults to '.'.}
}
\value{
TRUE on success, FALSE otherwise.
}
\description{
create directories of path '\code{path}/\code{scriptName}_type' and store
this path in an object called '\code{type}_directory'.
}
\note{
This functions creates the character vector \code{type}_directory with
content  '\code{path}/\code{scriptName}_type' in .GlobalEnv and the directory
'\code{path}/\code{scriptName}_type' on disk.
}
\examples{
provide_output_directory()
}
