% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wipe_clean.R
\name{wipe_clean}
\alias{wipe_clean}
\title{Remove All Objects From an Environment}
\usage{
wipe_clean(environment = .GlobalEnv)
}
\arguments{
\item{environment}{The environment that should be wiped clean.}
}
\value{
A character vector containing the names of objects removed.
}
\description{
Remove All Objects From an Environment
}
