% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cran.R
\name{install_cran}
\alias{install_cran}
\title{Forced Installation of CRAN Packages}
\usage{
install_cran(package = "excerptr", repository = "http://cran.r-project.org",
  ignore_ostype = TRUE, ignore_r_version = TRUE)
}
\arguments{
\item{repository}{The repository to use. Stick to the default, which is CRAN.}

\item{ignore_ostype}{Ignore operation system type required in the
DESCRIPTION?}

\item{ignore_r_version}{Ignore R version required in the DESCRIPTION?}
}
\value{
The value of \code{\link[utils:install.packages]{install.packages}}.
}
